Fix profiler url config
