+++ packages/dd-trace/src/profiling/config.js
@@ -2,7 +2,6 @@
 
 const coalesce = require('koalas')
 const os = require('os')
-const { URL } = require('url')
 const { AgentExporter } = require('./exporters/agent')
 const { FileExporter } = require('./exporters/file')
 const { ConsoleLogger } = require('./loggers/console')
@@ -15,10 +14,7 @@ const {
   DD_ENV,
   DD_TAGS,
   DD_SERVICE,
-  DD_VERSION,
-  DD_TRACE_AGENT_URL,
-  DD_AGENT_HOST,
-  DD_TRACE_AGENT_PORT
+  DD_VERSION
 } = process.env
 
 class Config {
@@ -44,13 +40,8 @@ class Config {
     this.logger = ensureLogger(options.logger)
     this.flushInterval = flushInterval
 
-    const hostname = coalesce(options.hostname, DD_AGENT_HOST, 'localhost')
-    const port = coalesce(options.port, DD_TRACE_AGENT_PORT, 8126)
-    const url = new URL(coalesce(options.url, DD_TRACE_AGENT_URL,
-      `http://${hostname || 'localhost'}:${port || 8126}`))
-
     this.exporters = ensureExporters(options.exporters || [
-      new AgentExporter({ url })
+      new AgentExporter(options)
     ], options)
 
     this.profilers = options.profilers || [
